src/requirements.txt
exif==1.6.0
ffmpeg-python==0.2.0
future==0.18.3
pillow==10.2.0
plum-py==0.8.7
